requirements.txt
altgraph==0.16.1
Django==2.1
django-crispy-forms==1.7.2
django-tinymce4-lite==1.7.1
future==0.16.0
html5lib==1.0.1
httplib2==0.11.3
jsmin==2.2.2
macholib==1.10
pefile==2018.8.8
Pillow==5.1.0
PyInstaller==3.3.1
PyPDF2==1.26.0
pypiwin32==223
pytz==2018.4
pywin32==223
reportlab==3.4.0
six==1.11.0
webencodings==0.5.1
xhtml2pdf==0.2.2



Django==2.0.2
django-extensions==1.9.9
django-tinymce4-lite==1.7.0
jsmin==2.2.2
Pillow==5.0.0
pkg-resources==0.0.0
pytz==2017.3
six==1.11.0
typing==3.6.4